List unit, dipisahkan perunit dan dibuat punya harga unit

+++ sysinvest/src/main/java/com/rj/sysinvest/model/Investment.java
@@ -40,7 +40,7 @@ public class Investment implements Serializable {
     private Timestamp marketRateUpdate;
 
     @Column(length = 19)
-    private Double soldRate;
+    private Long soldRate;
     
     @Column(nullable = true)
     private Timestamp soldTrx = null;
